refactor: share prototype variant switcher
